Move Bookeeping gerneration into generator.
