+++ lib/isogram_cases.rb
@@ -1,38 +1,5 @@
 require 'ostruct'
-
-class TestCase
-  attr_reader :data
-  def initialize(json_data)
-    @data = OpenStruct.new(json_data)
-  end
-
-  def method_definition
-    format 'def %s', test_name
-  end
-
-  def method_end
-    'end'
-  end
-
-  def test_name
-    format 'test_%s', @data['description'].downcase.tr_s(' -','_')
-  end
-
-  def workload
-    body = skip,
-      format( "string = '%s'", data.input ),
-      format( "%s Isogram.is_isogram?(string)", assertion)
-    [indent( body ),nil].join("\n")
-  end
-
-  def full_method
-    [indent([method_definition],2), workload.split("\n"), indent([method_end],2)].flatten.join("\n") + "\n"
-  end
-
-  def indent(array, count=4 )
-    array.map { |line| format "%s%s", ' ' * count, line }
-  end
-end
+require_relative 'testcase'
 
 class IsogramCase < TestCase
 
@@ -48,14 +15,13 @@ class IsogramCase < TestCase
     data.expected
   end
 
-
   def assertion
-    expected ? 'assert' : 'refute'
+    [
+      format( "string = '%s'", data.input ),
+      format( "%s Isogram.is_isogram?(string)", expected ? 'assert' : 'refute')
+    ]
   end
 
-  def skip
-    index.zero? ? '# skip' : 'skip'
-  end
 end
 
 IsogramCases = proc do |data|
